requirements.txt
python-dotenv==1.1.1
discord.py